move Traverse func to util.go - fix outer map not initialised in Reg struct
